Feature/conversation

+++ server/src/main/kotlin/com/vl/messenger/chat/ConversationService.kt
@@ -3,11 +3,12 @@ package com.vl.messenger.chat
 import com.vl.messenger.DataMapper
 import com.vl.messenger.PdfService
 import com.vl.messenger.StorageService
+import com.vl.messenger.asConversationDialogId
 import com.vl.messenger.profile.NotificationService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
-import java.io.OutputStream
+import java.io.ByteArrayOutputStream
 
 @Service
 class ConversationService(
@@ -16,25 +17,61 @@ class ConversationService(
     @Autowired private val pdfService: PdfService,
     @Autowired private val notificationService: NotificationService
 ) {
+    fun createConversation(userId: Int, name: String) =
+        dataMapper
+            .createConversation(userId, name)
+            .asConversationDialogId()
 
-    fun createConversation(userId: Int, name: String) = dataMapper.createConversation(userId, name)
+    fun setConversationName(userId: Int, conversationId: Long, name: String): CommonResult {
+        if (!hasPrivilege(userId, conversationId, Privilege.EDIT_DATA))
+            return CommonResult.NO_PRIVILEGE
 
-    fun getConversation(conversationId: Long) = dataMapper.getConversation(conversationId)
-
-    fun setConversationName(conversationId: Long, name: String) = dataMapper.setConversationName(conversationId, name)
-
-    fun setConversationImage(conversationId: Long, image: MultipartFile) {
-        val path = storageService.saveConversationImage(image, conversationId)
-        dataMapper.setConversationImage(conversationId, path)
+        dataMapper.setConversationName(conversationId, name)
+        return CommonResult.SUCCESS
     }
 
-    fun getMembers(conversationId: Long, offset: Int, limit: Int) = dataMapper.getMembers(conversationId, offset, limit)
+    fun setConversationImage(userId: Int, conversationId: Long, image: MultipartFile): CommonResult {
+        if (!hasPrivilege(userId, conversationId, Privilege.EDIT_DATA))
+            return CommonResult.NO_PRIVILEGE
+
+        val path = storageService.saveConversationImage(image, conversationId)
+        dataMapper.setConversationImage(conversationId, path)
+        return CommonResult.SUCCESS
+    }
 
     /**
-     * @param userId id of competent user having privilege to manage members
-     * @param memberId id of user being added
+     * Get role of currently authenticated user
      */
-    fun addMember(userId: Int, conversationId: Long, memberId: Int) { // TODO add members via invite rather than directly
+    fun getRole(userId: Int, conversationId: Long): CommonResultValue<DataMapper.ConversationMember.Role> =
+        dataMapper.getRole(userId, conversationId)
+            ?.let { CommonResultValue.Success(it) }
+            ?: CommonResultValue.NoPrivilege
+
+    /**
+     * Get roles specific for certain conversation
+     */
+    fun getRoles(userId: Int, conversationId: Long): CommonResultValue<List<DataMapper.ConversationMember.Role>> =
+        if (!hasPrivilege(userId, conversationId, Privilege.PARTICIPATE))
+            CommonResultValue.NoPrivilege
+        else
+            CommonResultValue.Success(dataMapper.roles) // roles are shared for all conversations yet
+
+    fun getMembers(userId: Int, conversationId: Long, offset: Int, limit: Int): GetMembersResult {
+        if (getConversation(conversationId) == null)
+            return GetMembersResult.NotFound
+
+        if (!hasPrivilege(userId, conversationId, Privilege.PARTICIPATE))
+            return GetMembersResult.NoPrivilege
+
+        return GetMembersResult.Success(
+            dataMapper.getMembers(conversationId, offset, limit)
+        )
+    }
+
+    fun addMember(userId: Int, conversationId: Long, memberId: Int): CommonResult {
+        if (!hasPrivilege(userId, conversationId, Privilege.EDIT_MEMBERS))
+            return CommonResult.NO_PRIVILEGE
+
         dataMapper.addMember(memberId, conversationId)
         notificationService.addNotification(
             memberId,
@@ -45,9 +82,13 @@ class ConversationService(
                 dataMapper.getConversation(conversationId)!!.name
             }"
         )
+        return CommonResult.SUCCESS
     }
 
-    fun removeMember(userId: Int, conversationId: Long, memberId: Int) {
+    fun removeMember(userId: Int, conversationId: Long, memberId: Int): CommonResult {
+        if (!hasPrivilege(userId, conversationId, Privilege.EDIT_MEMBERS))
+            return CommonResult.NO_PRIVILEGE
+
         dataMapper.removeMember(memberId, conversationId)
         notificationService.addNotification(
             memberId,
@@ -58,18 +99,29 @@ class ConversationService(
                 dataMapper.getConversation(conversationId)!!.name
             }"
         )
+        return CommonResult.SUCCESS
     }
 
     fun leaveConversation(userId: Int, conversationId: Long) =
         dataMapper.removeMember(userId, conversationId)
 
-    fun setMemberRole(conversationId: Long, memberId: Int, role: String) =
-        dataMapper.setRole(memberId, conversationId, role)
+    fun setMemberRole(userId: Int, conversationId: Long, memberId: Int, roleId: Int): SetRoleResult {
+        if (!hasPrivilege(userId, conversationId, Privilege.EDIT_RIGHTS))
+            return SetRoleResult.NO_PRIVILEGE
+
+        if (dataMapper.roles.firstOrNull { it.id == roleId } == null)
+            return SetRoleResult.ROLE_NOT_FOUND
+
+        dataMapper.setRole(memberId, conversationId, roleId)
+        return SetRoleResult.SUCCESS
+    }
 
     fun isMember(userId: Int, conversationId: Long) =
         dataMapper.getRole(userId, conversationId) != null
 
-    fun hasPrivilege(userId: Int, conversationId: Long, privilege: Privilege): Boolean {
+    private fun getConversation(conversationId: Long) = dataMapper.getConversation(conversationId)
+
+    private fun hasPrivilege(userId: Int, conversationId: Long, privilege: Privilege): Boolean {
         val role = dataMapper.getRole(userId, conversationId)
             ?: return false
         return when (privilege) {
@@ -92,6 +144,40 @@ class ConversationService(
     /**
      * Writes pdf file to output stream
      */
-    fun generateReport(conversationId: Long, outputStream: OutputStream): Unit =
-        pdfService.generateConversationActivityReport(dataMapper.getUsersActivity(conversationId), outputStream)
+    fun generateReport(userId: Int, conversationId: Long): GenerateReportResult {
+        if (!hasPrivilege(userId, conversationId, Privilege.GET_REPORTS))
+            return GenerateReportResult.NoPrivilege
+
+        return ByteArrayOutputStream().let {
+            pdfService.generateConversationActivityReport(dataMapper.getUsersActivity(conversationId), it)
+            GenerateReportResult.Success(it.toByteArray())
+        }
+    }
+
+    enum class CommonResult {
+        SUCCESS,
+        NO_PRIVILEGE
+    }
+
+    enum class SetRoleResult {
+        SUCCESS,
+        NO_PRIVILEGE,
+        ROLE_NOT_FOUND
+    }
+
+    sealed interface CommonResultValue<out T> {
+        data object NoPrivilege: CommonResultValue<Nothing>
+        data class Success<T>(val value: T): CommonResultValue<T>
+    }
+
+    sealed interface GetMembersResult {
+        data object NoPrivilege: GetMembersResult
+        data object NotFound: GetMembersResult
+        data class Success(val members: List<DataMapper.ConversationMember>): GetMembersResult
+    }
+
+    sealed interface GenerateReportResult {
+        data class Success(val reportFile: ByteArray): GenerateReportResult
+        data object NoPrivilege: GenerateReportResult
+    }
 }
